cse-v2-20190225T004945Z-001/cse-v2/requirements.txt
setuptools
wheel
requests
tzlocal
simple-crypt
python-ldap
mysqlclient
pymysql
python-dateutil
fpdf
python-Levenshtein
fuzzywuzzy
sqlalchemy
Flask
flask-classful
flask-login
Flask-SQLAlchemy
werkzeug
rauth
oauthlib
flask-dance
itsdangerous
blinker
gunicorn
